+++ server/Session.js
@@ -2,7 +2,7 @@ const { ALERT_TYPE } = require('brave-alert-lib')
 
 class Session {
   // prettier-ignore
-  constructor(id, chatbotState, alertType, numButtonPresses, createdAt, updatedAt, incidentCategory, buttonBatteryLevel, respondedAt, respondedByPhoneNumber, button) {
+  constructor(id, chatbotState, alertType, numButtonPresses, createdAt, updatedAt, incidentCategory, respondedAt, respondedByPhoneNumber, button) {
     this.id = id
     this.chatbotState = chatbotState
     this.alertType = alertType
@@ -10,7 +10,6 @@ class Session {
     this.createdAt = createdAt
     this.updatedAt = updatedAt
     this.incidentCategory = incidentCategory
-    this.buttonBatteryLevel = buttonBatteryLevel
     this.respondedAt = respondedAt
     this.respondedByPhoneNumber = respondedByPhoneNumber
     this.button = button
@@ -19,10 +18,6 @@ class Session {
   incrementButtonPresses(numButtonPresses) {
     this.numButtonPresses += numButtonPresses
     this.alertType = ALERT_TYPE.BUTTONS_URGENT
-  }
-
-  updateBatteryLevel(buttonBatteryLevel) {
-    this.buttonBatteryLevel = buttonBatteryLevel
   }
 }